+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor GUI to use Wails
+- Renamed:
+  - `SD Connect` to `SD-Connect` in the filesystem
+  - `SD Apply` to `SD-Apply` in the filesystem
+
+### Added
+
+- `-project` parameter, which can be used to override the SD Connect project in the VM
 
 ## [v1.4.1] - 2023-01-09